test: Move all run-make rustdoc tests to test/rustdoc
