+++ src/test/auxiliary/rustdoc-default-impl.rs
