+++ src/test/auxiliary/rustdoc-extern-default-method.rs
