+++ src/test/auxiliary/rustdoc-extern-method.rs
