+++ src/test/auxiliary/rustdoc-ffi.rs
