--- src/test/run-make/rustdoc-smoke/Makefile
@@ -1,4 +1,0 @@
--include ../tools.mk
-all:
-	$(HOST_RPATH_ENV) $(RUSTDOC) -w html -o $(TMPDIR)/doc foo.rs
-	$(HTMLDOCCK) $(TMPDIR)/doc foo.rs
