--- src/test/run-make/rustdoc-src-links/Makefile
@@ -1,5 +1,0 @@
--include ../tools.mk
-all:
-	$(HOST_RPATH_ENV) $(RUSTDOC) -w html -o $(TMPDIR)/doc foo.rs
-	$(HTMLDOCCK) $(TMPDIR)/doc foo.rs
-	$(HTMLDOCCK) $(TMPDIR)/doc qux/mod.rs
