+++ src/test/rustdoc/hidden-line.rs
@@ -8,8 +8,6 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-#![crate_name="foo"]
-
 /// The '# ' lines should be removed from the output, but the #[derive] should be
 /// retained.
 ///
@@ -31,5 +29,5 @@
 /// ```
 pub fn foo() {}
 
-// @!has foo/fn.foo.html invisible
+// @!has hidden_line/fn.foo.html invisible
 // @matches - //pre "#\[derive\(PartialEq\)\] // Bar"
