+++ src/test/rustdoc/recursion1.rs
