+++ src/test/rustdoc/recursion2.rs
