+++ src/test/rustdoc/recursion3.rs
