+++ src/test/rustdoc/search-index.rs
