+++ src/test/rustdoc/smoke.rs
@@ -8,29 +8,28 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-// @has foo/index.html
-#![crate_name = "foo"]
+// @has smoke/index.html
 
 //! Very docs
 
-// @has foo/bar/index.html
+// @has smoke/bar/index.html
 pub mod bar {
 
     /// So correct
-    // @has foo/bar/baz/index.html
+    // @has smoke/bar/baz/index.html
     pub mod baz {
         /// Much detail
-        // @has foo/bar/baz/fn.baz.html
+        // @has smoke/bar/baz/fn.baz.html
         pub fn baz() { }
     }
 
     /// *wow*
-    // @has foo/bar/trait.Doge.html
+    // @has smoke/bar/trait.Doge.html
     pub trait Doge { fn dummy(&self) { } }
 
-    // @has foo/bar/struct.Foo.html
+    // @has smoke/bar/struct.Foo.html
     pub struct Foo { x: isize, y: usize }
 
-    // @has foo/bar/fn.prawns.html
+    // @has smoke/bar/fn.prawns.html
     pub fn prawns((a, b): (isize, usize), Foo { x, y }: Foo) { }
 }
